+++ internal/pkg/utils/token/jwt.go
@@ -11,7 +11,7 @@ import (
 )
 
 type JwtPayload struct {
-	UserID   uint   `json:"user_id"`
+	UserID   int64   `json:"user_id"`
 	Username string `json:"username"`
 }
